Mark virtual machines as has customer data
At a certain stage in a virtual machine's lifecycle, it may become
tainted with user data.

From the moment we start the hypervisor for the first time, it's
uncertain when the user logged in and began processing. After this, the
virtual machine may have customer data if it wasn't created for a VM
pool. Virtual machines from the pool are considered to has some customer
data once they are selected from the pool. Customers cannot access them
until they are chosen from the pool. If the virtual machine hasn't been
marked as has_customer_data before its destroy, we mark it during the
destroy process.

When a virtual machine is considered might have some customer data, it
theoretically contains irreplaceable and private data.

We can develop various logics based on this value.

For instance, we discussed some 'recreate' logics for virtual machines.
We can easily recreate the virtual machines which don't have any
customer data, but we need to exercise caution with tainted ones.

Furthermore, this value can be utilized in our pool logic. We can select
only the virtual machines which don't have any customer data from the
pool.