Use has_customer_data value to pick a vm from the pool
We don't disassociate the virtual machines from the pool when picked
because we aim to maintain a fixed size of virtual machines provisioned
by the pool. We join the VM dataset with the runner table to identify
idle virtual machines, a practice that previously led to an incident.
This incident occurred when I modified the runner to not wait for the
destruction of the VM, resulting in virtual machines being selected by
two different runners.

In addition, we lock all virtual machines in the pool while picking one,
it inadvertently locks the picked virtual machine that's in use by a
runner, leading to deadlocks. We shouldn't lock the virtual machines
that have already been picked.

We can use the has_customer_data value to pick an idle virtual machine
from the pool. We can also use the updated row count to manage
concurrent virtual machine selections by multiple threads. This approach
allows us to avoid locking the virtual machines table entirely.