Improve  available storage accounting in DownloadBootImage
Previously, we ran LearnStorage after an image was downloaded to update
StorageDevice::available_storage_gib of storage devices on the host.

However, this caused problems when there were existing storage volumes
on the host: A storage volume's file can be sparse, in which case `df`
will report a lower occupied space for it. This causes problems down the
road.

This PR changes that so we instead deduct the downloaded image size from
available_storage_gib.