Use implied re-raising of exceptions
I think it is more idiomatic to only use `raise` when re-raising an
exception.  The use of the `raise expr` form indicates that something
more complex is going on (e.g. that a new exception might have been
bound to a variable), when in these cases, that is not true.