Upgrade to Ruby 3.2.4
When running rubocop, the new version of the parser gem complains
about ruby running with an older version of ruby, though it doesn't
cause any problem.  But, in general, we try to keep up to date on the
point releases anyhow.

Ruby 3.3 also has its first point release 3.3.1, it might be a good
time to upgrade.  After this.

The version of bundler I obtained with the new Ruby is likewise newer,
and our needs there are not particularly exotic, so, update it as
well.