Validate account names
The account name is used in the email templates such as project
invitations. If the account name isn't validated, it can be exploited to
insert harmful links into emails.  A user could alter their name to a
URL, enabling them to send email invitations with malicious links.

I've decided not to allow "." as well. This invalidates names like 'John
Jr. Doe', but services like Gmail make links clickable even without the
'http://' prefix. So, it's better to be safe than sorry.