Remove AlmaLinux 9.1 from boot image options
The URL that we had for AlmaLinux 9.1 previously was a general URL
without the 9.1 information which pointed to a AlmaLinux 9.3 download
when I checked it, so I renamed it to AlmaLinux 9.3 in the rhizome side.

So, currently if a customer requests AlmaLinux 9.1, we'll get paged.

I suggest we remove AlmaLinux from options customers can choose for now,
until we properly figure out how we should manage AlmaLinux images.