Refactor IPv6 generator to support hosts with smaller than prefix 64
We used to generate 2 bytes long random bits and shift it to the head of
guest prefix, assuming that these random bits will fit in between host's
and the guests' network. It is fine for hosts with /64 prefix. However,
there are providers who would assign the same /64 prefix to multiple
hosts, such as Leaseweb. In that case, this random ipv6 generator fails
to properly allocate an ipv6 prefix from the host's network.

To fix this issue, we reduce the random bit count to a single byte. This
way, if the host has a prefix length of less than 64 but larger than 72,
we can still allocate guest prefixes at the length of /80.

We are also fixing a bug that would cause a crash in case of the host
prefix is too long and bytes_needed returns 0. In that case, the unpack1
call would crash since SecureRandom.random_bytes(0) returns nil. We
simplify the whole function by simply generating a random_number in the
range of [2..255].