Migrations to add size_gib to the boot_image table
We need this when removing a boot image, when we want to update
available storage on the storage device.

Alternatively, we could get the size using an ssh command when removing
the boot image. But since this is also useful as a way of
double-checking approximate total used size on a storage device, where
we can just sum vm_storage_volume and boot_image sizes on it.