Program to remove a boot image.
Once we install a newer version of an image, we might want to remove the
older versions. This program enables that.

To use it, an operator is expect to run the following in the REPL:

```
> boot_image = BootImage[... id ...]
> st = boot_image.remove_boot_image
```

Then it'll wait for storage volumes using that boot image to drain and
after that it'll remove the image files and db records.