Do not select unversioned image if there is an active versioned image
When ordering by version, PostgreSQL defaults to returning NULLs first.
This implies that even if we download the most recent version of an
image, an available unversioned image will be chosen first. However, as
we're beginning to disallow unversioned images, we can consider them
outdated and avoid selecting them if an active versioned image exists.