Reinstate AlmaLinux image downloading with revised versioning
This is partially a revert of
69bf5c861656dc4423c5772ffd1747651661f307. But it also modifies the
versioning to be less specific: AlmaLinux is now requested only by its
major version, e.g. "9", not "9.3". It also adds almalinux-8, for
which we have some demand, and as it has similar looking URLs and
considerations I thought to include it.

This method is the most like Ubuntu even though it includes one less
version number component.  Ubuntu has releases like "22.04.3", where
we never displayed the last version component.  The "3" in AlmaLinux
"9.3" is a similar minor version that we don't need to display while
remaining consistent with Ubuntu.

That I included the second digit of AlmaLinux before was a mistake I
made in the most ancient era: March 2023, whereas the entire code base
began in January 2023, adc109d2ea570a170e71e339304c502b7d03c798.