Move default image versions from host_nexus to download_boot_image
We were expecting to pass image version to DownloadBootImage prog. Default
version of images were kept in host_nexus. With this commit default version
info has been moved to DownloadBootImage. If a version is not passed to
DownloadBootImage it will download the default version, if it is passed
given version will be downloaded.