Add github runner e2e tests
Adding e2e tests for github runners. It basically adds a new strand to
the e2e ci. That strand first downloads github runner images to test
machine, triggers the test worklfows passed in the github runner test
config file and then wait for them to be completed successfully.

Since we are running ngrok on the test host and set the ngrok's static
domain as the base uri for the test app we create, hooks of triggered
workflows will be sent to test host.

If you want to add a new repo for github runner tests, make sure that
the test app is installed for that repo.