Add active_from field to billing rates
We sometimes need to update billing rates, however, directly changing the rates
would affect the historical data. To avoid this, we are adding an active_from
field. Thanks to this field, we can create a new billing rate with the updated
data and more recent active_from date. The old billing rate will still be used
for the historical data and existing resources, but new resources will use the
new billing rate.