Add billing rates for VM storage
We used to have a single billing rate for VMs, which was combination of the
amount we would want to charge for both compute and storage. We are planning to
decouple storage and compute, thus we need to have separate billing rates for
each. This commit adds the billing rates for VM storage and also adds new rates
for VM compute, which is lower than the combined rate we used to have.