Add backend code allowing picking storage independently from VM size
We are expanding the VmSize option to include minimum and maximum storage sizes
as well as step size per each instance size. These will be passed to frontend
to dynamically update the user interface.

We are also adding storage_size option to API, which needs to be added to our
API documentation.