Add frontend code allowing picking storage independently from VM size
This frontend code is admittedly a bit of patch work on top of our existing UI
code. So far frontend code was not our primary focus and as a result of that,
it become a bit difficult to extend and test. I initially wanted to clean the
UI code up a bit, but that effort quickly exploded into a big refactor touching
lots of different areas, such as adding javascript tests, modifying the way we
render the forms, etc. I decided to postpone that work for now and make minimal
set of changes to get the storage decoupling feature working.

I'm planning to come back to form rendering part of the UI code in the future
and refactor it to make it more maintainable and testable.