Make Authorization.matched_policies accept polymorphic object_id
With this change, we will be able to use both id and ubid in matched_policies.
The motivation for this change is to be able to use ubids directly in our web
endpoints. If we use ubids, we can avoid passing the actual ids to frontend,
and eventually use same serializers for both web and api endpoints.