Inline serializers in roda branches
We used to define a global serializer at the top of the roda branch, and then
calls to `serialize` would use that serializer. This was a bit confusing for
few reasons;

1. In some cases, we use different serializers in same branch. To be able to
use second serializer, we need to explicitly call define which serializer to
use, however for the default case we don't need to do that. At a glance, it's
hard to understand which serializer would be used while resolving `serialize`
calls.
2. In most cases, we define global serializer but use it only once.

Overall, I think it is better to explicitly set the serializer at the time of
the serialization, not at the top of the branch.

As an extra benefit, this also reduces the line count by 42.