Allow serializers to be configured with options
We currently use structures for serializing same objects in different formats,
such as :default and :detailed. However, this approach is not flexible, because
to add an option which modifies both :default and :detailed formats, you need
to create 2 new structures.

For example, if you want to output slightly different information for :api and
:web options, you need to define :default_api, :detailed_api, :default_web and
:detailed_web structures.

With this change, you can pass options to the serializer, which then can be
used to alter the output. We also started to use options to distinguish between
default and detailed formats.

Apart from being more flexible, this approach allows using serializers without
creating serializer instances as all the methods become class methods. We also
cut the line count by 123.