Generalize safe writing to file
Create a new method `safe_create_file_with_content` that is similar to
`safe_write_to_file` but more flexible: Instead of taking a filename
together with the content as a string, `safe_create_file_with_content`
takes a filename and a block, which is the content producer.
`safe_write_to_file` is kept but now calls into
` safe_create_file_with_content`.