Create standbys in PostgresResource's wait state
We have very similar standby creation logic in two places: assemble and wait.
When we need to make a change on the standby creation logic, we need to update
two places. We can consolidate the standby creation logic into the wait state
since all creations would eventually reach there. One drawback of this is we
would create standbys ~2-3 minutes later than before, but it is a small time
window and it is at the beginning of the resource's lifecycle, hence not very
critical.