Cache and set the value of slider while changing its properties
Value in the slider might be invalid temporarily while changing min/max/step.
In such cases browser assigns a valid value automatically, which is not
necessarily the value we would pick. To prevent this, we read the value of the
slider before changing its properties and set it back after changing them.