Use vm.id while generating billing record for the IP address
We are grouping the billing records by the resource id while generating the
invoices. Thanks to this both compute and storage billing records of VMs are
grouped together. It makes sense to group the IP address billing records with
the VM billing records as well.

With this change, we also don't need to finalize the billing record through the
IP address, as the billing record will belong to VM and will be finalized as
part of VM's billing record finalization. However for the existing records
still needs to be finalized through the IP address. In the future, when there
is no such billing record left, we can remove the finalize call for the IP
address.