Introduce GitHubCacheEntry model and add blob to repository
The cache entry contains metadata about the available caches for a
specific repository. The unique identifiers for the cache, `key` and
`version`, are used by the action client to either retrieve or upload a
cache. The `scope` aids in maintaining the branch where the cache was
created, as each branch cannot access the cache of another. The
`upload_id` is generated when creating a multipart upload request on
blob storage, which is then used to upload cache parts and complete the
upload. Additionally, we keep track of the cache's last access time to
clear old caches that haven't been accessed for a considerable period.

Each repository has its unique cache, necessitating a separate bucket
for each repository. This separation naturally demarcates the caches of
different repositories. Services like S3 and MinIO, which are S3
compatible, allow for the creation of access policies based on the
object key prefix. Thus, we can limit access to a repository's cache by
using the repository name as the object key prefix within the same
bucket. However, Cloudflare R2 doesn't currently support this feature,
requiring us to create a separate bucket and access token for each
repository.

Since these credentials can only access the repository's bucket, we use
them for object-level operations like generating presigned URLs and
managing multipart uploads.

url_presigner helps to generate presigned URLs. We don't do any call to
genereate resigned URLs, it's generatable on the client side using its
credentials.