Add base of GitHub runtime API
The VM runtime token is a JWT token used for authenticating virtual
machines with our runtime API, and it remains valid while the VM is
running. We utilize the runner assigned to the VM and the repository
associated with this runner in the "/runtime/github" endpoints.
Subsequent commits introduce these new endpoints.

We pass the runtime token and URL to the runner as environment variables
during the setup process.