Add getCacheEntry endpoint
We've implemented caching endpoints based on GitHub's official cache
client. We also have a forked version, which gives us flexibility, but
we've tried to stay as close as possible to the official client,
maintaining similar endpoints and parameters.

The official GitHub cache client is open-source and can be accessed
here [^1]

The 'GET /cache' endpoint ('getCacheEntry' [^2]) retrieves cache entry using
comma-separated 'keys' and 'version', returning a pre-signed URL to
download the cache. If no cache entry is found, it returns a 204 status
code.

Therefore, careful selection of cache entries is crucial.

Initially, we only consider committed caches for the repository of the
authenticated runner. The runner can only access caches for the same
branch or the default branch, but not for other branches.

The client can send multiple keys, and we should return the first
matching cache entry in the same order as the keys are provided. The
`.min_by { keys.index(_1.key) }` function assists with this. The
`(scopes.index(_1.scope) * keys.size)` part helps us sort the caches by
scope. If the same cache exists for the head_branch and the default
branch, we should return the cache for the head_branch.

[^1]: https://github.com/actions/toolkit/blob/main/packages/cache/src/internal/cacheHttpClient.ts
[^2]: https://github.com/ubicloud/toolkit/blob/d1df13e178816d69d96bdc5c753b36a66ad03728/packages/cache/src/internal/cacheHttpClient.ts#L103