Add reserveCache endpoint
The client reserves a cache entry before beginning to upload content to
the blob storage. S3 compatible blob storage necessitates sending a
`CreateMultipartUpload` request prior to the content upload. This
returns an `UploadId`.

The 'POST /caches' endpoint, ('reserveCache' [^1]), creates a cache
entry using the given "key" and "version" parameters. This prevents any
other runner from starting to upload the same cache. It returns an
`upload_id` from the blob storage and a list of presigned URLs for each
32MB part, which the client will use to upload each chunk.

The cache client uses the `cacheId` in some logics. This is unnecessary
for our forked version since we already have `uploadId`. However, to
avoid extensive modifications to the client code, a fixed `cacheId` is
returned for each request.

[^1]: https://github.com/ubicloud/toolkit/blob/d1df13e178816d69d96bdc5c753b36a66ad03728/packages/cache/src/internal/cacheHttpClient.ts#L204