Add commitCache endpoint
The client commit the cache entry after uploading content to the blob
storage. S3 compatible blob storage requires sending a
`CompleteMultipartUpload` request following the content upload. This
necessitates an `UploadId` and `ETags` from each part of the upload.

The 'POST /caches/commit' endpoint ('commitCache' [^1]) sends a
`CompleteMultipartUpload` request to the blob storage using the etags
provided by the client. Afterward, it marks the cache entry as
committed, allowing other runners to access it.

[^1]: https://github.com/ubicloud/toolkit/blob/d1df13e178816d69d96bdc5c753b36a66ad03728/packages/cache/src/internal/cacheHttpClient.ts#L339