Persist default branch information for GitHub repositories
We need the default branch information of a GitHub repository to
implement certain features. For instance, GitHub Action Cache
integration requires all other branches to access the main branch's
cache. Continually fetching the default branch isn't feasible when we
need it. As the workflow_job webhook event payload already includes the
default branch information and we run insert_conflict query for the
repository, we also store the default branch information in the
repository table.