Use storage_size_options array instead of min/max/step values
For specifying the storage size options for a given instance type for VM and
Postgres resources, we used to pass the min, max and step values. This has few
problems:

- We cannot use options with varying step values
- We had to generate the options array in few places which was repetitive

With this commit, we are starting to generate storage_size_options array and
use it in the views and validations.