Clean up invalid CSRF exception handling
The `parse_error` in `clover_base` expects the
`Roda::RodaPlugins::RouteCsrf::InvalidToken` to be defined. This is true
if the `route_csrf` plugin is enabled, but it might not be in some
cases. Only `CloverWeb` uses CSRF protection and has an additional check
for it. Thus, we can combine them and pass it during the plugin
configuration.