Add base class for common routes
We are working towards combining business logic in web and api endpoints. For
that purpose, we will introduce helper classes that will be shared between the
two routes. This is the base class for those helper classes.

The base class is quite simple at the moment. It only includes an initializer
and few helper methods to easily access app object's fields.