Add helper class for postgres routes
The business logic in both web and api endpoints are combined in the helper
class. Currently the combination is made simply adding a branch for each option
and copying the business logic from each endpoint to their respective branches.
Rubocop also moved some common lines out of the branches, but other than that
we did not try to converge single logic and eliminate the branches as much as
possible. This is intentional as I wanted to keep this commit as boring and add
a separate commit with the more interesting convergence stuff.

In its current form, this commit does not provide much benefit and in fact it
increases the total line count. However it paves the way for making complex
improvements.