Add retry logic for processes in E2E tests
We realized that in some cases, respirate exits due to a crash or apoptosis,
which causes foreman to exit and E2E tests to fail. Foreman is programmed to
exit when one of its child processes exits. It doesn't have any retry logic.
Recommended way of adding retry logic is combining it with a monitoring tool
such as systemd or upstart. However, we don't need such a complex system for
our E2E tests. Instead I opted to add retry logic with simple `until` loop.

I didn't add the retry logic for `bin/ci` process we are adding in next step
of the workflow file, because tests themselves might not be written in a way
that allows them to be executed second time.