Make VM E2E tests compatible with AlmaLinux images
We've made AlmaLinux 8 and AlmaLinux 9 available to our customers.
AlmaLinux uses the dnf package manager, not apt. The changes in this PR
ensure that our E2E tests are compatible with AlmaLinux images.

The command `sudo dnf groupinstall -y 'Development Tools'` is similar to
`sudo apt install -y build-essential`, which we use for Ubuntu images.
However, the AlmaLinux version is significantly larger, so we've opted
for a lighter version instead.

`sudo dnf check-update` exits with 100 if there are available package
updates. Since non-zero exit codes indicate failure in bash world, I
added `|| [ $? -eq 100 ]` condition to suppress false error if it's 100.