Handle the deletion of the runner with running job smoother
Before we destroy the underlying virtual machine of the runner, we must
ensure that it is not executing any job and has been deleted on GitHub's
side. This precaution helps us avoid disrupting potentially active jobs.

Ideally, when we initiate the runner's deletion, it should have already
completed its job. However, in some instances, such as when the runner
script fails due to out-of-memory (OOM) errors, network issues, etc., it
loses communication with GitHub. In such cases, GitHub delays marking
the runner as idle to rule out intermittent issues. If the runner is
still executing a job when we attempt its deletion, it raises an
'Octokit::UnprocessableEntity' exception. Instead of raising this
exception and logging the entire stack trace, we can simply nap for a
while and retry the deletion.

Additionally, I have extended the deadline from 10 to 15 minutes.
Lately, we've encountered this situation due to some OOM errors.
Typically, these are resolved between 10-15 minutes. There's no need for
additional page.