Download boot images from blob storage for E2E tests
There are several benefits of downloading boot images from blob storage:

- It is faster as most of the data transfer would be done in same region or in
the network of same provider.
- It is more reliable because there is no risk of source to be deleted or moved
without our knowledge.

The motivation trigger for this change was to reduce the time taken to download
the boot images, because our E2E tests are timing out due to slow download, but
the change is useful even out of E2E tests. So currently, I enabled this only
for E2E tests, but we can enable this for production as well in the future. We
would probably keep existing way for development because it doesn't require
setting up blob storage, thus it keeps the development setup simple.