Use arch data from label directly instead of an extra branch
We don't need to use the label string to identify the runner's
architecture, as this information is already in the label data. We
reference this data in several places within this file. I've extracted
it to a global variable. Additionally, removing this branch eliminated
the need for an extra test to cover it.