Lock the lock file before opening the temp file while writing to a file
If two concurrent processes try to write to the same file, the second one would
be able to open the temp file and would wait until the first one finishes, but
the first one renames the temp file, which leaves the second process in a weird
state. We fixed this problem in control plane's safe_write_to_file method, but
we didn't apply the fix in the data plane version of it. This commit fixes the
problem in the data plane version of the method by locking a lock file before
opening the temp file.