Use more idiomatic(?) style for raising a new exception in one case
I inherited opinions about this rather than having my own, but upon
closer scrutiny, this really doesn't matter, because `raise` and
`fail` have identical semantics.

For some reason, I always thought `fail` *would not* do implicit
re-raise (but it does), and this is why `fail` was preferred for
de-novo exceptions and `raise` was preferred in some other contexts.
But it's simply not true.

It's worth thinking if we should teach rubocop to rewrite all `fail`
as `raise` by adding a rule.  Or vice versa if we liked being special
and saving a character.