UI: Default to location with highest available capacity
Before this change, during VM creation, the first element of of the
"Location" radio button is selected by default.  After this change,
the location with the highest available capacity is selected.

"Available capacity" is determined as the sum of available cores in a
location. There are other options to calculate/interpret available
capacity. The main drawback of this approach is that it doesn't consider
fragmentation. Another option would be to pick the location with the
host that has the max available capacity.  But in that case we'd
currently default to our smallest location (US South).