Add OS users to be used by the prometheus integration
We added 2 new OS users; prometheus and ubi_monitoring, which are created while
generating the base PostgreSQL OS image. The prometheus user will be used to
run the prometheus server and node_exporter and the former means that it will
be exposed to outside. The ubi_monitoring user will be used to connect to the
database and scrape Postgres metrics. We are adding a special row for allowing
the ubi_monitoring OS user to connect to the database as the ubi_monitoring
database user. This user has pg_monitor rights, thus it is able to read/execute
various monitoring views and functions. I didn't want to allow prometheus to
connect to the database because it is exposed to the outside and potential
vulnerability on the prometheus server could lead to the database compromise
otherwise.