Configure TLS for prometheus web server
Prometheus exposes a web server, which both provides a primitive UI and also
endpoints for scraping metrics. This commit configures the web server to use
TLS, which is done by providing a certificate and key file. We are reusing the
certificate we generated for the postgres as both postgres and prometheus
servers are exposed from the same URL, only the ports will differ.