Allow to specify default firewall rules for private subnets
When a private subnet is created, it also creates a firewall under the
hood. This firewall is created with a default set of rules that allow
all traffic. This is not always desirable, so this commit allows to
specify the default rules when creating a private subnet.

The call to create a private subnet without firewall rules still works
and continues to create a default firewall with "allow all" rules.

curl "https:/private-subnet/subnet-name" \
  -X POST \
  -H "Authorization: Bearer $jwt_token"

But now it's also possible to specify the rules of the default firewall
when creating the subnet, like this:

curl "https:/private-subnet/subnet-name" \
  -X POST \
  -H "Authorization: Bearer $jwt_token" \
  -H "Content-Type: application/json" \
  -d '{
      "firewall_rules": [
			{
				"cidr": "10.0.0.0/24",
				"port_range": "22"
			},
			{
				"cidr": "0.0.0.0/0",
				"port_range": "5432..5434"
			}
		]
	}'

This creates a private subnet with a default firewall that denies
all incoming traffic:

curl "https:/private-subnet/subnet-name" \
  -X POST \
  -H "Authorization: Bearer $jwt_token" \
  -H "Content-Type: application/json" \
  -d '{
      "firewall_rules": []
	}'