Allow to specify default firewall rules for a vm
When a VM is created without a private network, a private subnet is
created and the VM is attached to it. This subnet is created with a
default firewall that allows all traffic. This commit allows to specify
the firewall rules when creating a VM.

Omitting firewall rules will continue to create a default subnet with
a firewall that allows all traffic.

curl "https://.../vm/vm-name" \
  -X POST \
  -H "Authorization: Bearer $jwt_token" \
  -H "Content-Type: application/json" \
  -d '{
	    "public_key": "the public ssh key"
	}'

But now it's possible to specify the firewall rules:

curl "https://.../vm/vm-name" \
  -X POST \
  -H "Authorization: Bearer $jwt_token" \
  -H "Content-Type: application/json" \
  -d '{
	    "public_key": "the public ssh key",
      "firewall_rules": [
			{
				"cidr": "10.0.0.0/24",
				"port_range": "22"
			},
			{
				"cidr": "0.0.0.0/0",
				"port_range": "5432..5434"
			},
			{
				"cidr": "10.1.2.3/32"
			}
		]
	}'

It is also still possible to specify the private subnet to attach the VM
to, in which case no additional firewalls or firewall rules are created:

curl "https://.../vm/vm-name" \
  -X POST \
  -H "Authorization: Bearer $jwt_token" \
  -H "Content-Type: application/json" \
  -d '{
	    "public_key": "the public ssh key",
	    "private_subnet_id": "the ps id"
	}'