Configure basic auth for prometheus web server
Prometheus exposes a web server that can be used to query metrics. This web
server accepts unauthenticated requests by default. This commit adds basic
authentication to the web server. We use the ubid of the PostgresResource as
the username and the password is a random string that is generated when the
resource is created.