Reduce respirate sleep time from 5s to 1s
We have observed that respirate frequently enters the sleep
state, particularly during periods of lower traffic. When new
"work" arrives just after respirate begins sleeping, it can introduce
a delay of up to 5 seconds. This delay is relatively long, so we
are reducing the sleep time to 1 second.

During periods of high traffic, respirate does not sleep often, making
the sleep time less relevant. Additionally, sleeping for extended
periods during high traffic is undesirable as it increases the risk of
falling behind.