Assemble private subnet with an existing firewall
When creating a private subnet, a new firewall is typically generated
and linked to the subnet. This commit allows for the option of
specifying a firewall_id. By providing an existing firewall_id, no new
firewall will be created. Instead, the specified existing firewall will
be attached to the newly created private subnet.