Send invoice failure notification to accounts with billing permissions
When an invoice fails to process, we send an email to the billing email
address. Our experience shows that customers often provide service email
addresses that aren't frequently checked. Consequently, the response
rate from these emails is lower than that from users' registered email
addresses. We will now also send an email to accounts with billing
permissions for the project.