Remove flowtables from the veth interface
Keeping flowtables for the veth interface causes routing failure for
consecutive packages. After the initial packet hits the prerouting chain
and perform the dnat, we want it to continue with the post routing and
then resubmitted immediately to the kernel for ESP encapsulation.
However, with flowtable, the packets are immediately transformed to
their pre ESP phase (having private addresses in source and destination)
and then submitted back to veth interface. This causes packets to leak
out to vetho* interface, therefore, we miss the chance of creating an
ESP packet. This caused failures for balancing, therefore, we limit the
flowtables to the tap* device.