Allow passing multiple ssh keys while setting up VM
We used to allow passing only one ssh key while setting up a VM. For the VMs
that belongs to services, it would be useful to pass multiple ssh keys. For
example, we can pass array of operator keys to be put into the VMs.

An alternative approach would be adding SSH keys to the VM using the sshable
interface. However, that would happen after certain steps in the service setup,
thus wouldn't be useful for debugging bootstrapping issues. It would also
require each service to implement key addition logic.